[UPDATE]버그 수정 + Report API 붙이기 + 폰트/백버튼 일괄 적용

+++ GitCat/GitCat/Main/Util/Model/ReportDetailModel.swift
@@ -28,7 +28,7 @@ class LanguageRatioModel:Codable {
     var resultLanguages:[ResultLanguagesModel?]
 }
 class ContributedRepositoryModel:Codable {
-    var count:[Int?]
+    var count:[Double?]
     var repoNames:[String?]
 }
 class ResultLanguagesModel:Codable {
